Update db-create-connection and move transferRepo
