--- website/test-balance/src/domain/gear/gear.service.ts
@@ -1,106 +1,0 @@
-// import { GearApi, GearKeyring, TransferData } from '@gear-js/api';
-// import { initLogger } from '@gear-js/common';
-// import { BN } from '@polkadot/util';
-// import { KeyringPair } from '@polkadot/keyring/types';
-//
-// import config from '../../config/configuration';
-// import { ResponseTransferBalance } from '../../gear/types';
-// import { transferService } from '../transfer/transfer.service';
-//
-// const log = initLogger('GearService');
-//
-// export class GearService {
-//   api: GearApi;
-//   account: KeyringPair;
-//   rootAccount: KeyringPair;
-//   accountBalance: BN;
-//   balanceToTransfer: BN;
-//
-//   constructor() {}
-//
-//   public async connect() {
-//     this.api = await GearApi.create({ providerAddress: config.gear.providerAddress });
-//     this.api.on('error', () => {
-//       GearApi.create({ providerAddress: config.gear.providerAddress }).then(
-//         (newApi) => {
-//           this.api = newApi;
-//         },
-//         (error) => {
-//           console.error('could not reconnect:', error);
-//           throw error;
-//         },
-//       );
-//     });
-//     log.info(`Connected to ${await this.api.chain()} with genesis ${this.getGenesisHash}`);
-//
-//     const [account, rootAccountByGearKeyring] = await Promise.all([
-//       GearKeyring.fromSeed(config.gear.accountSeed),
-//       this.setRootAccountSeedByGearKeyring(),
-//     ]);
-//
-//     this.account = account;
-//     this.rootAccount = rootAccountByGearKeyring;
-//     this.accountBalance = new BN(config.gear.accountBalance);
-//     this.balanceToTransfer = new BN(config.gear.balanceToTransfer);
-//     if (await this.isSmallAccountBalance()) {
-//       await this.transferBalance(this.account.address, this.rootAccount, this.accountBalance);
-//     }
-//   }
-//
-//   public async transferBalance(
-//     to: string,
-//     from: KeyringPair = this.account,
-//     balance: BN = this.accountBalance,
-//   ): Promise<ResponseTransferBalance> {
-//     try {
-//       await this.transfer(from, to, balance);
-//     } catch (error) {
-//       log.error(error);
-//       return { error: `Transfer balance from ${from} to ${to} failed` };
-//     }
-//     if (to !== this.account.address) {
-//       await transferService.setTransferDate(to, this.getGenesisHash);
-//     }
-//     return { status: 'ok', transferredBalance: balance.toString() };
-//   }
-//
-//   public get getGenesisHash(): string {
-//     return this.api.genesisHash.toHex();
-//   }
-//
-//   private async transfer(
-//     from: KeyringPair = this.account,
-//     to: string,
-//     balance: BN = this.accountBalance,
-//   ): Promise<TransferData> {
-//     this.api.balance.transfer(to, balance);
-//     return new Promise((resolve, reject) => {
-//       this.api.balance.signAndSend(from, ({ events }) => {
-//         events.forEach(({ event: { method, data } }) => {
-//           if (method === 'Transfer') {
-//             resolve(data as TransferData);
-//           } else if (method === 'ExtrinsicFailed') {
-//             reject(data);
-//           }
-//         });
-//       });
-//     });
-//   }
-//
-//   private async isSmallAccountBalance(): Promise<boolean> {
-//     const balance = await this.api.balance.findOut(this.account.address);
-//     if (balance.lt(this.accountBalance)) {
-//       return true;
-//     }
-//     return false;
-//   }
-//
-//   private async setRootAccountSeedByGearKeyring(): Promise<KeyringPair> {
-//     const envVar = config.gear.rootAccountSeed;
-//     if (envVar === '//Alice') {
-//       return GearKeyring.fromSuri('//Alice');
-//     } else {
-//       return GearKeyring.fromSeed(config.gear.rootAccountSeed);
-//     }
-//   }
-// }
